+++ adapters/sigfox-vertx/src/main/java/org/eclipse/hono/adapter/sigfox/impl/SigfoxProtocolAdapter.java
@@ -154,25 +154,26 @@ public final class SigfoxProtocolAdapter
 
         final String deviceId = ctx.getRoutingContext().queryParams().get(SIGFOX_PARAM_DEVICE_ID);
         final String strData = ctx.getRoutingContext().queryParams().get(SIGFOX_PARAM_DATA);
-        final Buffer data = decodeData(strData);
 
         LOG.debug("{} handler - deviceTenant: {}, requestTenant: {}, deviceId: {}, data: {}",
                 ctx.request().method(), deviceTenant, requestTenant, deviceId, strData);
 
-        if ( requestTenant == null ) {
-            ctx.fail(new ClientErrorException(HttpURLConnection.HTTP_BAD_REQUEST,
+        if (requestTenant == null) {
+            ctx.fail(new ClientErrorException(HttpURLConnection.HTTP_NOT_FOUND,
                     "missing the tenant information in the request URL"));
             return;
         }
 
         if (!requestTenant.equals(deviceTenant)) {
-            ctx.fail(new ClientErrorException(HttpURLConnection.HTTP_BAD_REQUEST,
+            ctx.fail(new ClientErrorException(HttpURLConnection.HTTP_FORBIDDEN,
                     "tenant information mismatch"));
             return;
         }
 
-        final String contentType = (data != null) ? CONTENT_TYPE_OCTET_STREAM
-                : EventConstants.CONTENT_TYPE_EMPTY_NOTIFICATION;
+        final Buffer data = decodeData(strData);
+        final String contentType = Optional.ofNullable(data)
+                .map(d -> CONTENT_TYPE_OCTET_STREAM)
+                .orElse(EventConstants.CONTENT_TYPE_EMPTY_NOTIFICATION);
 
         uploadHandler.upload(ctx, deviceTenant, deviceId, data, contentType);
     }
@@ -195,7 +196,7 @@ public final class SigfoxProtocolAdapter
 
     private static Buffer decodeData(final String data) {
         if (data == null) {
-            return Buffer.buffer();
+            return null;
         }
         return Buffer.buffer(BaseEncoding.base16().decode(data.toUpperCase()));
     }